add and edit multiple validators
