+++ src/Validation/BCH.php
@@ -2,10 +2,10 @@
 
 namespace Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
 
-use Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
+use Merkeleon\PhpCryptocurrencyAddressValidation\Base58Validation;
 use Merkeleon\PhpCryptocurrencyAddressValidation\Utils\CashAddress;
 
-class BCH extends Validation
+class BCH extends Base58Validation
 {
     // more info at https://en.bitcoin.it/wiki/List_of_address_prefixes
     protected $base58PrefixToHexVersion = [
