+++ src/Validation/BSV.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
+
+
+class BSV extends BCH
+{
+
+}
