+++ src/Validation/ETC.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
+
+class ETC extends ETH
+{
+
+}
