+++ src/Validation/NEO.php
@@ -2,14 +2,13 @@
 
 namespace Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
 
-
 use Merkeleon\PhpCryptocurrencyAddressValidation\Base58Validation;
 
-class DOGE extends Base58Validation
+class NEO extends Base58Validation
 {
+    // more info at https://en.bitcoin.it/wiki/List_of_address_prefixes
     protected $base58PrefixToHexVersion = [
-        'D' => '1E',
-        '9' => '16',
-        'A' => '16',
+        'A' => '17'
     ];
+
 }
