+++ src/Validation/TBTC.php
@@ -6,12 +6,13 @@ use Merkeleon\PhpCryptocurrencyAddressValidation\Base58Validation;
 use Merkeleon\PhpCryptocurrencyAddressValidation\Utils\Bech32Decoder;
 use Merkeleon\PhpCryptocurrencyAddressValidation\Utils\Bech32Exception;
 
-class BTC extends Base58Validation
+class TBTC extends Base58Validation
 {
     // more info at https://en.bitcoin.it/wiki/List_of_address_prefixes
     protected $base58PrefixToHexVersion = [
-        '1' => '00',
-        '3' => '05'
+        'm' => '6F',
+        'n' => '6F',
+        '2' => 'C4'
     ];
 
     public function validate($address)
@@ -21,7 +22,7 @@ class BTC extends Base58Validation
         if (!$valid) {
             // maybe it's a bech32 address
             try {
-                $valid = is_array($decoded = Bech32Decoder::decodeRaw($address)) && 'bc' === $decoded[0];
+                $valid = is_array($decoded = Bech32Decoder::decodeRaw($address)) && 'tb' === $decoded[0];
             } catch (Bech32Exception $exception) {}
         }
