+++ src/Validation/ZEC.php
@@ -2,9 +2,9 @@
 
 namespace Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
 
-use Merkeleon\PhpCryptocurrencyAddressValidation\Validation;
+use Merkeleon\PhpCryptocurrencyAddressValidation\Base58Validation;
 
-class ZEC extends Validation
+class ZEC extends Base58Validation
 {
     // more info at https://en.bitcoin.it/wiki/List_of_address_prefixes
     protected $base58PrefixToHexVersion = [
